Moved unit tests for GetDecksByDescriptionContaining() into DeckRepositoryTests.
