+++ AnkiJapaneseFlashcardManagerTest/AnkiJapaneseFlashcardManagerTests.csproj
@@ -57,8 +57,4 @@
     </None>
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="DataAccessLayer\Repositories\" />
-  </ItemGroup>
-
 </Project>
